PR-04c: extract kanata.rs (KanataClient + protocol + ShutdownGuard)
Behaviour-preserving relocation of the kanata TCP client. Highest-
coupling extraction so far — every backend consumes KanataClient.

kanata.rs (new): all *Msg/*Payload wire types, KanataClientInner,
KanataClient + full impl, ShutdownGuard + Drop. KanataClient stays
`pub` per plan lock; everything else `pub(crate)`. Whitebox tests
required widening the inner field and three helper fns to pub(crate).

main.rs: 6588 -> 5945 LOC.

Verification: cargo build clean; cargo test --bin kanata-switcher
261 passed / 0 failed.
